signal_engine/requirements.txt
numpy==1.24.3
pandas==2.0.3
ta==0.11.0
ccxt==4.2.25
requests==2.31.0
python-telegram-bot==20.7
aiohttp==3.9.1
uvloop==0.19.0
joblib==1.3.2
scikit-learn==1.3.2
python-dotenv==1.0.0
asyncio==3.4.3
